Improved the ChemObject flags API. We can now directly clear/set a flag without having a branch (if/else). Likewise due to padding a 16-bit integer does not take less space than 32-bit in this case and the 32-bit int is cleaner to work with.

+++ base/test/src/test/java/org/openscience/cdk/test/interfaces/AbstractChemObjectTest.java
@@ -159,14 +159,14 @@ public abstract class AbstractChemObjectTest extends AbstractCDKObjectTest {
     @Test
     public void testGetFlagValueZeroDefault() {
         IChemObject chemObject = newChemObject();
-        Assertions.assertEquals((short) 0, chemObject.getFlagValue());
+        Assertions.assertEquals((int) 0, chemObject.getFlagValue());
     }
 
     @Test
     public void testGetFlagValue() {
         IChemObject chemObject = newChemObject();
         chemObject.setFlag(CDKConstants.ISALIPHATIC, true);
-        Assertions.assertNotSame((short) 0, chemObject.getFlagValue());
+        Assertions.assertNotSame((int) 0, chemObject.getFlagValue());
     }
 
     /**
